fix code climate
