+++ pkg/gofr/datasource/pubsub/mqtt/mqtt_pub.go
@@ -1,0 +1,45 @@
+package mqtt
+
+import (
+	"context"
+	"time"
+
+	"go.opentelemetry.io/otel"
+
+	"gofr.dev/pkg/gofr/datasource/pubsub"
+)
+
+func (m *MQTT) Publish(ctx context.Context, topic string, message []byte) error {
+	_, span := otel.GetTracerProvider().Tracer("gofr").Start(ctx, "mqtt-publish")
+	defer span.End()
+
+	m.metrics.IncrementCounter(ctx, "app_pubsub_publish_total_count", "topic", topic)
+
+	s := time.Now()
+
+	token := m.Client.Publish(topic, m.config.QoS, m.config.RetrieveRetained, message)
+
+	// Check for errors during publishing (More on error reporting
+	// https://pkg.go.dev/github.com/eclipse/paho.mqtt.golang#readme-error-handling)
+	if token.Wait() && token.Error() != nil {
+		m.logger.Errorf("error while publishing message, error: %v", token.Error())
+
+		return token.Error()
+	}
+
+	t := time.Since(s)
+
+	m.logger.Debug(&pubsub.Log{
+		Mode:          "PUB",
+		CorrelationID: span.SpanContext().TraceID().String(),
+		MessageValue:  string(message),
+		Topic:         topic,
+		Host:          m.config.Hostname,
+		PubSubBackend: "MQTT",
+		Time:          t.Microseconds(),
+	})
+
+	m.metrics.IncrementCounter(ctx, "app_pubsub_publish_success_count", "topic", topic)
+
+	return nil
+}
